requirements.txt
ccxt
pandas
polars
pyarrow
duckdb
tqdm
tenacity
python-dateutil
requests
vectorbt
pandas_ta
pyyaml
aiohttp
httpx
feedparser
trafilatura
aiolimiter
tldextract
simhash
langdetect
pydantic
web3
eth-abi
hexbytes
python-dotenv
